Add beginning of dedup, POST first
Deduping is a reutilization of ideas behind bup and camlistore: content
is chunked around content-defined boundaries (there's a good read at
https://github.com/bup/bup/blob/master/DESIGN, the splitting is the
same except we don't build trees and we don't care about all the git
stuff, only chunks and a root entity), the individual chunks are
stored as-is, and the chunklist is stored as the content of the "file".
For reading we have the checklist and reconstruct the full content based
on that.